// apps/indexer/src/database/schemas/public/Nft.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import { type ColumnType, type Selectable, type Insertable, type Updateable } from 'kysely';

/** Identifier type for public.Nft */
export type NftId = string & { __brand: 'NftId' };

/** Represents the table public.Nft */
export default interface NftTable {
  id: ColumnType<NftId, NftId | undefined, NftId>;

  collectionAddress: ColumnType<string, string, string>;

  address: ColumnType<string | null, string | null, string | null>;

  tokenId: ColumnType<bigint | null, bigint | null, bigint | null>;

  nftIndex: ColumnType<bigint, bigint, bigint>;

  name: ColumnType<string, string, string>;

  image: ColumnType<string, string, string>;

  description: ColumnType<string, string, string>;

  uri: ColumnType<string, string, string>;

  raw: ColumnType<unknown, unknown, unknown>;
}

export type Nft = Selectable<NftTable>;

export type NewNft = Insertable<NftTable>;

export type NftUpdate = Updateable<NftTable>;